requirements.txt
absl-py==2.0.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Flask==3.0.0
Flask-HTTPAuth==4.8.0
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.2
pandas==1.5.2
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stem==1.8.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
torrequest==0.1.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.9.0
urllib3==1.26.13
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
